Wordle solver fully running
This whole program should work now. It asks you for your first guess, then after that it will instruct you what to guess. It still is very weak when it gets down to the last 10ish words. It is also very weak against words that end in similar letters to other words such as: "abel","inds","est"